feat: bcs-task add task builder
